+++ src/OrdersList/OrdersListFiltersContainer.test.js
@@ -2,6 +2,10 @@ import { render } from '@folio/jest-config-stripes/testing-library/react';
 
 import OrdersListFiltersContainer from './OrdersListFiltersContainer';
 
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
+  useFunds: jest.fn(() => ({ funds: [{ id: 'id', code: 'code' }] })),
+}));
 jest.mock('./OrdersListFilters', () => jest.fn().mockReturnValue('OrdersListFilters'));
 
 const defaultProps = {